% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alra.R
\name{ALRA}
\alias{ALRA}
\title{ALRA}
\usage{
ALRA(object, assay_use = "RNA", slot_use = "data", ...)
}
\arguments{
\item{object}{Data object to impute}

\item{assay_use}{The assay from which to retrieve data. Default: "RNA"}

\item{slot_use}{The slot within the assay to impute. Default: "data"}

\item{...}{additional parameters to pass to ALRA::alra()}
}
\description{
Use Adaptively-thresholded Low Rank Approximation to perform imputation
}
