% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoImpute.R
\name{autoimpute}
\alias{autoimpute}
\title{AutoImpute}
\usage{
autoimpute(object, assay = "RNA", slot.use = "data",
  assay.store = "RNA", slot.store = "data", normalize.data = FALSE,
  scale.data = TRUE, debug = TRUE, debug_display_step = 1,
  hidden_units = 2000, lambda_val = 1, initial_learning_rate = 1e-04,
  iterations = 7000, threshold = 1e-04, masked_matrix_test = FALSE,
  masking_percentage = 10, log_file = "log.txt", load_saved = FALSE)
}
\arguments{
\item{debug}{}

\item{debug_display_step}{}

\item{hidden_units}{}

\item{lambda_val}{}

\item{initial_learning_rate}{}

\item{iterations}{}

\item{threshold}{}

\item{masked_matrix_test}{}

\item{masking_percentage}{}

\item{log_file}{}

\item{load_saved}{}

\item{exprData}{}
}
\description{
AutoImpute is an auto-encoder based gene-expression (sparse) matrix imputation.
(https://github.com/divyanshu-talwar/AutoImpute)
Some arguments have been removed due to being troublesome in the R-Python
interface
}
