% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherData.R
\docType{methods}
\name{GatherData,Seurat-method}
\alias{GatherData,Seurat-method}
\title{GatherData}
\usage{
\S4method{GatherData}{Seurat}(object, assay = "RNA", slot.use = "data",
  ...)
}
\arguments{
\item{object}{Seurat object to pull data from}

\item{assay}{(Default: "RNA")}

\item{slot.use}{(Default: "data)}

\item{...}{}
}
\value{
matrix
}
\description{
GatherData
}
