% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{GatherData.Seurat}
\alias{GatherData.Seurat}
\alias{GatherData.seurat}
\alias{GatherData.SingleCellExperiment}
\title{GatherData}
\usage{
\method{GatherData}{Seurat}(object, assay = "RNA", slot_use = "counts",
  ...)

\method{GatherData}{seurat}(object, slot_use = "data", ...)

\method{GatherData}{SingleCellExperiment}(object,
  assay_use = "logcounts", ...)
}
\arguments{
\item{object}{Seurat object to pull data from}

\item{assay}{(Default: "RNA")}

\item{slot_use}{(Default: "data)}

\item{...}{}
}
\value{
matrix

matrix

matrix
}
\description{
GatherData
}
