% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{Magic}
\alias{Magic}
\alias{Magic.default}
\alias{Magic.Seurat}
\alias{Magic.seurat}
\alias{Magic.SingleCellExperiment}
\title{Magic}
\usage{
Magic(object, ...)

\method{Magic}{default}(exprDat, genes = NULL, k = 10, alpha = 15,
  t = "auto", npca = 100, init = NULL, t.max = 20,
  knn.dist.method = "euclidean", verbose = 1, n.jobs = 1,
  seed = NULL)

\method{Magic}{Seurat}(object, assay_use = "RNA", slot_use = "counts",
  assay_store = "magic", genes = NULL, k = 10, alpha = 15,
  t = "auto", npca = 100, init = NULL, t.max = 20,
  knn.dist.method = "euclidean", verbose = 1, n.jobs = 1,
  seed = NULL)

\method{Magic}{seurat}(object, slot_use = "data",
  assay_store = "magic", genes = NULL, k = 10, alpha = 15,
  t = "auto", npca = 100, init = NULL, t.max = 20,
  knn.dist.method = "euclidean", verbose = 1, n.jobs = 1,
  seed = NULL)

\method{Magic}{SingleCellExperiment}(object, assay_use = "logcounts",
  assay_store = "magic", genes = NULL, k = 10, alpha = 15,
  t = "auto", npca = 100, init = NULL, t.max = 20,
  knn.dist.method = "euclidean", verbose = 1, n.jobs = 1,
  seed = NULL)
}
\arguments{
\item{object}{Seurat object with data to impute by MAGIC.}

\item{...}{Additional arguments}

\item{exprDat}{expression matrix to impute}

\item{genes}{Default: NULL}

\item{k}{Default: 10}

\item{alpha}{Default: 15}

\item{t}{Default: "auto"}

\item{npca}{Default: 10}

\item{init}{Default: NULL}

\item{t.max}{Default: 20}

\item{knn.dist.method}{Default: "euclidean"}

\item{verbose}{Default: 1}

\item{n.jobs}{Default: 1}

\item{seed}{Default: NULL}

\item{assay_use}{Object assay containing the data to impute. Default: "RNA" or "logcounts"}

\item{slot_use}{Assay slot containing data to impute. Default: "counts"}

\item{assay_store}{Name to use when storing imputed data. Default: "magic"}
}
\value{
matrix

Seurat object

seurat object

SingleCellExperiment object
}
\description{
Wrapper around the Markov Affinity-based Graph Imputation of Cells (MAGIC)
imputation algorithm from the KrishnaswamyLab
}
