% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaceData.R
\docType{methods}
\name{PlaceData,Seurat-method}
\alias{PlaceData,Seurat-method}
\title{PlaceData}
\usage{
\S4method{PlaceData}{Seurat}(object, assay = "new_assay",
  slot.use = "data", imputed_data, normalize = "FALSE",
  scale = "FALSE", ...)
}
\arguments{
\item{object}{Seurat object to pull data from}

\item{assay}{Name of assay slot to place imputated data}

\item{...}{}
}
\value{
matrix
}
\description{
PlaceData
}
