% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{PlaceData.Seurat}
\alias{PlaceData.Seurat}
\alias{PlaceData.seurat}
\alias{PlaceData.SingleCellExperiment}
\title{PlaceData}
\usage{
\method{PlaceData}{Seurat}(object, assay_store = "new_assay",
  imputed_data, normalize = "FALSE", scale = "FALSE", ...)

\method{PlaceData}{seurat}(object, assay_store = "new_assay",
  slot_use = "data", imputed_data, normalize = "FALSE",
  scale = "FALSE", ...)

\method{PlaceData}{SingleCellExperiment}(object,
  assay_store = "new_assay", imputed_data, normalize = "FALSE", ...)
}
\arguments{
\item{object}{Seurat object to pull data from}

\item{...}{}

\item{assay}{Name of assay slot to place imputated data}
}
\description{
PlaceData
}
