% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{RetrieveIdents}
\alias{RetrieveIdents}
\alias{RetrieveIdents.Seurat}
\alias{RetrieveIdents.seurat}
\alias{RetrieveIdents.SingleCellExperiment}
\title{RetrieveIdents}
\usage{
RetrieveIdents(object, ...)

\method{RetrieveIdents}{Seurat}(object, identity = NULL, ...)

\method{RetrieveIdents}{seurat}(object, identity = NULL, ...)

\method{RetrieveIdents}{SingleCellExperiment}(object, identity = NULL,
  ...)
}
\arguments{
\item{object}{Data object from which to retrieve identiy information}

\item{...}{Additional arguments to pass to child functions}

\item{identity}{Variable to use for grouping identity}
}
\value{
A list corresponding to the identities of each cell.

matrix

matrix

matrix
}
\description{
Helper function to retrive cell identity information
}
