% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimpute.R
\name{scImpute}
\alias{scImpute}
\title{scImpute}
\usage{
scImpute(object, assay = "RNA", slot = "counts", num_clusters = NULL,
  ident_use = NULL, ensembl_db = NULL, ...)
}
\arguments{
\item{object}{Data object to impute}

\item{assay}{Assay from which to retrieve expression data. Default: "RNA"}

\item{slot}{Slot from which to retrieve expression data. Default: "counts"}

\item{num_clusters}{Number of clusters to use when imputing data.  If left NULL,
will attempt to retrive the current identity information and use a count equal to the
number of unique identities.}

\item{ident_use}{Identity variable to use when imputing the data.}

\item{ensembl_db}{The Ensembl annotation package to use to compute gene lengths.
Default: EnsDb.Hsapiens.v86}

\item{...}{Additional arguments to pass to \code{\link{scImpute::scimpute}}}
}
\value{
Seurat
}
\description{
Impute data using scImpute
}
