% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{GatherData}
\alias{GatherData}
\alias{GatherData.Seurat}
\alias{GatherData.seurat}
\alias{GatherData.SingleCellExperiment}
\title{GatherData}
\usage{
GatherData(object, ...)

\method{GatherData}{Seurat}(object, assay = "RNA", slot_use = "counts")

\method{GatherData}{seurat}(object, assay = "RNA", slot_use = "data")

\method{GatherData}{SingleCellExperiment}(object, assay = "logcounts")
}
\arguments{
\item{object}{Seurat object to pull data from}

\item{...}{Additional arguments}

\item{assay}{(Default: "RNA")}

\item{slot_use}{(Default: "data)}
}
\value{
matrix

matrix

matrix
}
\description{
GatherData
}
