% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{PlaceData,Seurat-method}
\alias{PlaceData,Seurat-method}
\title{PlaceData}
\usage{
\S4method{PlaceData}{Seurat}(object, assay.store = "new_assay",
  imputed_data, normalize = "FALSE", scale = "FALSE", ...)
}
\arguments{
\item{object}{Seurat object to pull data from}

\item{...}{}

\item{assay}{Name of assay slot to place imputated data}
}
\value{
matrix
}
\description{
PlaceData
}
