% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{PlaceData}
\alias{PlaceData}
\alias{PlaceData.Seurat}
\alias{PlaceData.seurat}
\alias{PlaceData.SingleCellExperiment}
\title{PlaceData}
\usage{
PlaceData(object, ...)

\method{PlaceData}{Seurat}(object, assay_store = "new_assay",
  imputed_data, normalize = "FALSE", scale = "FALSE")

\method{PlaceData}{seurat}(object, assay_store = "new_assay",
  slot_use = "data", imputed_data, normalize = "FALSE",
  scale = "FALSE")

\method{PlaceData}{SingleCellExperiment}(object,
  assay_store = "new_assay", imputed_data, normalize = "FALSE")
}
\arguments{
\item{object}{Seurat object to pull data from}

\item{...}{Additional arguments}

\item{assay_store}{Name of assay slot to place imputated data}

\item{imputed_data}{Imputed matrix}

\item{normalize}{Should the imputed data be normalized?}

\item{scale}{Should the imputed data be scaled?}
}
\description{
PlaceData
}
